Prevent selecting values that don't lie on a step (#582)
This prevents from the input "settling" on a value that doesn't lie on a step.

Example, given a fan with step size 33.3333:

1. Fan is initially off
2. User drags or tap position ~40%
3. Slider "snaps" to position 33% (fan is set to 33%, aka low) - *this is expected, existing behavior in mushroom*
4. User again drags or taps again to position ~40%, slider will not "snap back" to the 33% mark - *this PR aims to fix that behavior*

In my opinion, I don't believe the slider show a value that is not represented by the entity. Going from 33% to 40% doesn't speed up the fan -- the value of 33% is sent to Home Assistant both times.

This behavior is most visible on fans, and admittedly I don't have any other entities that I can test this with. But I assume the same expectation exists for all stepped entities (fans, covers, etc.)