Update lock.md
hi,
tnx for your awsome work !

this changes a typo in the lock documentation..


the horizontal layout splits the tile into 2 areas.  it might be better to reserve more space for the info.
i don't know where to start to even alter this.    (could provide screenshot.)


thank you for this repository,
seppe peeters    peeterss83@gmail.com